// src/client/blog/api.ts
export interface CategoryPublic {
  /**
   *
   * @type {string}
   * @memberof CategoryPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof CategoryPublic
   */
  name: string
  /**
   *
   * @type {Array<SubcategoryPublic>}
   * @memberof CategoryPublic
   */
  subcategories?: Array<SubcategoryPublic>
}
// ...
export interface PostCreate {
  /**
   *
   * @type {string}
   * @memberof PostCreate
   */
  title?: string | null
  /**
   *
   * @type {string}
   * @memberof PostCreate
   */
  content?: string | null
  /**
   *
   * @type {boolean}
   * @memberof PostCreate
   */
  is_public?: boolean | null
  /**
   *
   * @type {string}
   * @memberof PostCreate
   */
  category_id?: string | null
  /**
   *
   * @type {string}
   * @memberof PostCreate
   */
  subcategory_id?: string | null
}
/**
 *
 * @export
 * @interface PostPublic
 */
export interface PostPublic {
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  title: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  content: string
  /**
   *
   * @type {boolean}
   * @memberof PostPublic
   */
  is_public: boolean
  /**
   *
   * @type {CategoryPublic}
   * @memberof PostPublic
   */
  category?: CategoryPublic | null
  /**
   *
   * @type {SubcategoryPublic}
   * @memberof PostPublic
   */
  subcategory?: SubcategoryPublic | null
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  author: string
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  created_at?: string | null
  /**
   *
   * @type {string}
   * @memberof PostPublic
   */
  updated_at?: string | null
}
/**
 *
 * @export
 * @interface PostUpdate
 */
export interface PostUpdate {
  /**
   *
   * @type {string}
   * @memberof PostUpdate
   */
  id: string
  /**
   *
   * @type {string}
   * @memberof PostUpdate
   */
  title?: string | null
  /**
   *
   * @type {string}
   * @memberof PostUpdate
   */
  content?: string | null
  /**
   *
   * @type {boolean}
   * @memberof PostUpdate
   */
  is_public?: boolean | null
  /**
   *
   * @type {string}
   * @memberof PostUpdate
   */
  category_id?: string | null
  /**
   *
   * @type {string}
   * @memberof PostUpdate
   */
  subcategory_id?: string | null
}
/**
 *
 * @export
 * @interface SubcategoryPublic
 */
export interface SubcategoryPublic {
  /**
   *
   * @type {string}
   * @memberof SubcategoryPublic
   */
  _id: string
  /**
   *
   * @type {string}
   * @memberof SubcategoryPublic
   */
  name: string
}